+++ rollup.config.mjs
@@ -1,10 +1,11 @@
+import { babel } from '@rollup/plugin-babel';
 import commonjs from '@rollup/plugin-commonjs';
 import nodeResolve from '@rollup/plugin-node-resolve';
-import typescript from '@rollup/plugin-typescript';
-import { babel } from '@rollup/plugin-babel';
 import terser from '@rollup/plugin-terser';
-import filesize from 'rollup-plugin-filesize';
+import typescript from '@rollup/plugin-typescript';
 import fs from 'fs';
+import process from 'process';
+import filesize from 'rollup-plugin-filesize';
 
 const babelConfig = {
   babelHelpers: 'bundled',
@@ -33,13 +34,20 @@ const componentNames = fs
   // 带上package/src/index.ts
   .concat({ path: 'index', name: 'index' });
 
+// 获取包名(当前所在的文件夹名)并转换成大驼峰
+const UMDName = process
+  .cwd()
+  .split('/')
+  .pop()
+  .replace(/(^|-)(\w)/g, (_, _$1, $2) => $2.toUpperCase());
+
 export default [
   {
     input: './src/index.ts',
     output: {
       dir: './umd',
       format: 'umd',
-      name: 'umd',
+      name: UMDName,
     },
     plugins: [
       typescript({